add confident score

+++ Dataset.py
@@ -301,7 +301,7 @@ class Dataset(Dataset):
                     previous_nodes += coco_body_point_num if self.is_coco else halpe_body_point_num
                     previous_nodes += head_point_num if index_body == 2 else 0
                 edge_index = torch.tensor(np.array(l_pair) - np.full((len(l_pair), 2), previous_nodes),
-                                          dtype=torch.long).t().contiguous()
+                                          dtype=torch.float16).t().contiguous()
                 x_tensor, e_tensor = torch.zeros(
                     (self.sample_fps * self.video_len * int(input_size / 3), 3)), torch.zeros(
                     (2, self.sample_fps * self.video_len * len(l_pair)))
